+++ python/dolfinx/wrappers/mesh.cpp
@@ -4,11 +4,11 @@
 //
 // SPDX-License-Identifier:    LGPL-3.0-or-later
 
+#include "mesh.h"
 #include "MPICommWrapper.h"
 #include "array.h"
 #include "caster_mpi.h"
 #include "numpy_dtype.h"
-#include <cfloat>
 #include <dolfinx/common/IndexMap.h>
 #include <dolfinx/fem/CoordinateElement.h>
 #include <dolfinx/fem/ElementDofLayout.h>
@@ -21,7 +21,6 @@
 #include <dolfinx/mesh/graphbuild.h>
 #include <dolfinx/mesh/topologycomputation.h>
 #include <dolfinx/mesh/utils.h>
-#include <iostream>
 #include <memory>
 #include <nanobind/nanobind.h>
 #include <nanobind/ndarray.h>
@@ -35,75 +34,6 @@
 #include <span>
 
 namespace nb = nanobind;
-
-namespace
-{
-/// Wrap a Python graph partitioning function as a C++ function
-template <typename Functor>
-auto create_partitioner_cpp(Functor p)
-{
-  return [p](MPI_Comm comm, int nparts,
-             const dolfinx::graph::AdjacencyList<std::int64_t>& local_graph,
-             bool ghosting)
-  {
-    return p(dolfinx_wrappers::MPICommWrapper(comm), nparts, local_graph,
-             ghosting);
-  };
-}
-
-/// Wrap a C++ cell partitioning function as a Python function
-template <typename Functor>
-auto create_cell_partitioner_py(Functor p)
-{
-  return [p](dolfinx_wrappers::MPICommWrapper comm, int n,
-             const std::vector<dolfinx::mesh::CellType>& cell_types,
-             std::vector<nb::ndarray<const std::int64_t, nb::numpy>> cells_nb)
-  {
-    std::vector<std::span<const std::int64_t>> cells;
-    std::ranges::transform(
-        cells_nb, std::back_inserter(cells), [](auto c)
-        { return std::span<const std::int64_t>(c.data(), c.size()); });
-    return p(comm.get(), n, cell_types, cells);
-  };
-}
-
-using PythonCellPartitionFunction
-    = std::function<dolfinx::graph::AdjacencyList<std::int32_t>(
-        dolfinx_wrappers::MPICommWrapper, int,
-        const std::vector<dolfinx::mesh::CellType>&,
-        std::vector<nb::ndarray<const std::int64_t, nb::numpy>>)>;
-
-using CppCellPartitionFunction
-    = std::function<dolfinx::graph::AdjacencyList<std::int32_t>(
-        MPI_Comm, int, const std::vector<dolfinx::mesh::CellType>& q,
-        const std::vector<std::span<const std::int64_t>>&)>;
-
-/// Wrap a Python cell graph partitioning function as a C++ function
-CppCellPartitionFunction
-create_cell_partitioner_cpp(const PythonCellPartitionFunction& p)
-{
-  if (p)
-  {
-    return [p](MPI_Comm comm, int n,
-               const std::vector<dolfinx::mesh::CellType>& cell_types,
-               const std::vector<std::span<const std::int64_t>>& cells)
-    {
-      std::vector<nb::ndarray<const std::int64_t, nb::numpy>> cells_nb;
-      std::ranges::transform(
-          cells, std::back_inserter(cells_nb),
-          [](auto c)
-          {
-            return nb::ndarray<const std::int64_t, nb::numpy>(
-                c.data(), {c.size()}, nb::handle());
-          });
-
-      return p(dolfinx_wrappers::MPICommWrapper(comm), n, cell_types, cells_nb);
-    };
-  }
-  else
-    return nullptr;
-}
-} // namespace
 
 namespace dolfinx_wrappers
 {
@@ -260,7 +190,8 @@ void declare_mesh(nb::module_& m, std::string type)
           "__init__",
           [](dolfinx::mesh::Mesh<T>* mesh, MPICommWrapper comm,
              std::shared_ptr<dolfinx::mesh::Topology> topology,
-             dolfinx::mesh::Geometry<T>& geometry) {
+             dolfinx::mesh::Geometry<T>& geometry)
+          {
             new (mesh) dolfinx::mesh::Mesh<T>(comm.get(), topology, geometry);
           },
           nb::arg("comm"), nb::arg("topology"), nb::arg("geometry"))
@@ -280,10 +211,11 @@ void declare_mesh(nb::module_& m, std::string type)
       create_interval.c_str(),
       [](MPICommWrapper comm, std::int64_t n, std::array<T, 2> p,
          dolfinx::mesh::GhostMode ghost_mode,
-         const PythonCellPartitionFunction& part)
+         const part::impl::PythonCellPartitionFunction& part)
       {
         return dolfinx::mesh::create_interval<T>(
-            comm.get(), n, p, ghost_mode, create_cell_partitioner_cpp(part));
+            comm.get(), n, p, ghost_mode,
+            part::impl::create_cell_partitioner_cpp(part));
       },
       nb::arg("comm"), nb::arg("n"), nb::arg("p"), nb::arg("ghost_mode"),
       nb::arg("partitioner").none());
@@ -293,12 +225,12 @@ void declare_mesh(nb::module_& m, std::string type)
       create_rectangle.c_str(),
       [](MPICommWrapper comm, std::array<std::array<T, 2>, 2> p,
          std::array<std::int64_t, 2> n, dolfinx::mesh::CellType celltype,
-         const PythonCellPartitionFunction& part,
+         const part::impl::PythonCellPartitionFunction& part,
          dolfinx::mesh::DiagonalType diagonal)
       {
         return dolfinx::mesh::create_rectangle<T>(
-            comm.get(), p, n, celltype, create_cell_partitioner_cpp(part),
-            diagonal);
+            comm.get(), p, n, celltype,
+            part::impl::create_cell_partitioner_cpp(part), diagonal);
       },
       nb::arg("comm"), nb::arg("p"), nb::arg("n"), nb::arg("celltype"),
       nb::arg("partitioner").none(), nb::arg("diagonal"));
@@ -308,11 +240,12 @@ void declare_mesh(nb::module_& m, std::string type)
       create_box.c_str(),
       [](MPICommWrapper comm, std::array<std::array<T, 3>, 2> p,
          std::array<std::int64_t, 3> n, dolfinx::mesh::CellType celltype,
-         const PythonCellPartitionFunction& part)
+         const part::impl::PythonCellPartitionFunction& part)
       {
         MPI_Comm _comm = comm.get();
-        return dolfinx::mesh::create_box<T>(_comm, _comm, p, n, celltype,
-                                            create_cell_partitioner_cpp(part));
+        return dolfinx::mesh::create_box<T>(
+            _comm, _comm, p, n, celltype,
+            part::impl::create_cell_partitioner_cpp(part));
       },
       nb::arg("comm"), nb::arg("p"), nb::arg("n"), nb::arg("celltype"),
       nb::arg("partitioner").none());
@@ -323,7 +256,7 @@ void declare_mesh(nb::module_& m, std::string type)
                                          nb::c_contig>>& cells_nb,
            const std::vector<dolfinx::fem::CoordinateElement<T>>& elements,
            nb::ndarray<const T, nb::c_contig> x,
-           const PythonCellPartitionFunction& p)
+           const part::impl::PythonCellPartitionFunction& p)
         {
           std::size_t shape1 = x.ndim() == 1 ? 1 : x.shape(1);
 
@@ -365,7 +298,7 @@ void declare_mesh(nb::module_& m, std::string type)
          nb::ndarray<const std::int64_t, nb::ndim<2>, nb::c_contig> cells,
          const dolfinx::fem::CoordinateElement<T>& element,
          nb::ndarray<const T, nb::c_contig> x,
-         const PythonCellPartitionFunction& p)
+         const part::impl::PythonCellPartitionFunction& p)
       {
         std::size_t shape1 = x.ndim() == 1 ? 1 : x.shape(1);
         if (p)
@@ -599,7 +532,8 @@ void mesh(nb::module_& m)
   m.def(
       "compute_entities",
       [](MPICommWrapper comm, const dolfinx::mesh::Topology& topology, int dim,
-         int index) {
+         int index)
+      {
         return dolfinx::mesh::compute_entities(comm.get(), topology, dim,
                                                index);
       },
@@ -733,9 +667,9 @@ void mesh(nb::module_& m)
 
   m.def(
       "create_cell_partitioner",
-      [](dolfinx::mesh::GhostMode gm) -> PythonCellPartitionFunction
+      [](dolfinx::mesh::GhostMode gm) -> part::impl::PythonCellPartitionFunction
       {
-        return create_cell_partitioner_py(
+        return part::impl::create_cell_partitioner_py(
             dolfinx::mesh::create_cell_partitioner(gm));
       },
       "Create default cell partitioner.");
@@ -746,11 +680,12 @@ void mesh(nb::module_& m)
              const dolfinx::graph::AdjacencyList<std::int64_t>& local_graph,
              bool ghosting)>
              part,
-         dolfinx::mesh::GhostMode ghost_mode) -> PythonCellPartitionFunction
+         dolfinx::mesh::GhostMode ghost_mode)
+          -> part::impl::PythonCellPartitionFunction
       {
-        return create_cell_partitioner_py(
+        return part::impl::create_cell_partitioner_py(
             dolfinx::mesh::create_cell_partitioner(
-                ghost_mode, create_partitioner_cpp(part)));
+                ghost_mode, part::impl::create_partitioner_cpp(part)));
       },
       nb::arg("part"), nb::arg("ghost_mode") = dolfinx::mesh::GhostMode::none,
       "Create a cell partitioner from a graph partitioning function.");
